modified for fp32 openbookqa

+++ collie/models/llama.py
@@ -515,7 +515,8 @@ def load_model(
     tokenizer = Tokenizer(model_path=tokenizer_path)
 
     model_args.vocab_size = tokenizer.n_words
-    torch.set_default_tensor_type(torch.cuda.BFloat16Tensor)
+    # torch.set_default_tensor_type(torch.cuda.BFloat16Tensor)
+    torch.set_default_tensor_type(torch.cuda.FloatTensor)
     #torch.set_default_tensor_type(torch.cuda.HalfTensor)
     model = Transformer(model_args)
     torch.set_default_tensor_type(torch.cuda.FloatTensor)
